Implement a feather atlas for msaa
Diffs=
8f7120837b Implement a feather atlas for msaa (#9001)

+++ renderer/src/shaders/glsl.glsl
@@ -441,7 +441,7 @@
 #    ifdef @ENABLE_SPIRV_CROSS_BASE_INSTANCE
        // This uniform is specifically named "SPIRV_Cross_BaseInstance" for compatibility with
        // SPIRV-Cross sytems that search for it by name.
-       uniform int $SPIRV_Cross_BaseInstance;
+       uniform highp int $SPIRV_Cross_BaseInstance;
 #      define INSTANCE_INDEX (gl_InstanceID + $SPIRV_Cross_BaseInstance)
 #    else
 #        define INSTANCE_INDEX (gl_InstanceID + gl_BaseInstance)
